requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-dotenv
requests
google-api-python-client
openai
google-auth 
google-auth-oauthlib 
google-auth-httplib2